requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
greenlet==3.0.3
idna==3.6
mysqlclient==2.2.1
pip==23.3.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==58.1.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
djangorestframework-simplejwt==5.3.1
pytest==7.4.4


